비회원 댓글 수정 가능하도록 & 비회원 댓글 삭제 전 비밀번호 창까지

+++ src/main/java/egovframework/com/cop/cmt/web/ArticleCommentController.java
@@ -1,11 +1,13 @@
 package egovframework.com.cop.cmt.web;
 
+import java.io.PrintWriter;
 import java.util.Enumeration;
 import java.util.HashMap;
 import java.util.Map;
 
 import javax.annotation.Resource;
 import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -21,6 +23,9 @@ import egovframework.com.cmm.LoginVO;
 import egovframework.com.cmm.util.EgovUserDetailsHelper;
 import egovframework.com.cop.cmt.service.Comment;
 import egovframework.com.cop.cmt.service.CommentVO;
+import egovframework.com.cop.bbs.service.BBSMasterService;
+import egovframework.com.cop.bbs.service.BoardMasterVO;
+import egovframework.com.cop.bbs.service.BoardVO;
 import egovframework.com.cop.cmt.service.ArticleCommentService;
 import egovframework.com.utl.fcc.service.EgovStringUtil;
 import egovframework.rte.fdl.property.EgovPropertyService;
@@ -55,6 +60,9 @@ public class ArticleCommentController {
     @Resource(name="egovMessageSource")
     EgovMessageSource egovMessageSource;
     
+	@Resource(name = "BBSMasterService")
+	private BBSMasterService BBSMasterService;
+    
     @Autowired
     private DefaultBeanValidator beanValidator;
     
@@ -69,7 +77,7 @@ public class ArticleCommentController {
      * @throws Exception
      */
     @RequestMapping("/cop/cmt/selectArticleCommentList.do")
-    public String selectArticleCommentList(@ModelAttribute("searchVO") CommentVO commentVO, ModelMap model, HttpServletRequest request) throws Exception {
+    public String selectArticleCommentList(@ModelAttribute("searchVO") CommentVO commentVO, BoardVO boardVO, ModelMap model, HttpServletRequest request) throws Exception {
 
     	CommentVO articleCommentVO = new CommentVO();
     	
@@ -87,17 +95,26 @@ public class ArticleCommentController {
 		LoginVO user = (LoginVO)EgovUserDetailsHelper.getAuthenticatedUser();
    	 	// KISA 보안취약점 조치 (2018-12-10, 신용호)
         Boolean isAuthenticated = EgovUserDetailsHelper.isAuthenticated();
+    	
+        BoardMasterVO master = new BoardMasterVO();
 
-        if(!isAuthenticated) {
-        	return "forward:/uat/uia/LoginUsr.do";
-        }
+		master.setBbsId(boardVO.getBbsId());
+		master.setUniqId((user == null || user.getUniqId() == null) ? "" : user.getUniqId());
+
+		BoardMasterVO masterVo = BBSMasterService.selectBBSMasterInf(master);
+		
+		if(!masterVo.getAuthRead().equals("GUE")) { // 접근 권한 비회원 이상(회원 & 관리자)			
+			if (!isAuthenticated ) { // 로그인이 안되어 있을 경우
+				return "forward:/uat/uia/LoginUsr.do";
+			}
+			if(!masterVo.getAuthRead().equals(user.getUserSe()) && !user.getUserSe().equals("USR")) { // 관리자가 아니고 회원 권한이 맞지 않을 경우
+				return "egovframework/com/admin/cmm/error/accessDenied";
+			}
+		}
 		
 		model.addAttribute("sessionUniqId", user == null ? "" : EgovStringUtil.isNullToString(user.getUniqId()));
 		
 		commentVO.setWrterNm(user == null ? "" : EgovStringUtil.isNullToString(user.getName()));
-		
-//		commentVO.setSubPageUnit(propertyService.getInt("pageUnit"));
-//		commentVO.setSubPageSize(propertyService.getInt("pageSize"));
 	
 		PaginationInfo paginationInfo = new PaginationInfo();
 		paginationInfo.setCurrentPageNo(commentVO.getSubPageIndex());
@@ -118,8 +135,7 @@ public class ArticleCommentController {
 		model.addAttribute("paginationInfo", paginationInfo);
 		model.addAttribute("type", "body");	// 댓글 페이지 body import용
 		
-		model.addAttribute("articleCommentVO", articleCommentVO);	// validator 용도 
-		model.addAttribute("user", user);	//추가
+		model.addAttribute("articleCommentVO", articleCommentVO);	// validator 용도 	
 		
 		commentVO.setCommentCn("");	// 등록 후 댓글 내용 처리
 		
@@ -145,7 +161,7 @@ public class ArticleCommentController {
      * @throws Exception
      */
     @RequestMapping("/cop/cmt/insertArticleComment.do")
-    public String insertArticleComment(@ModelAttribute("searchVO") CommentVO commentVO, @ModelAttribute("comment") Comment comment, 
+    public String insertArticleComment(@ModelAttribute("searchVO") CommentVO commentVO, @ModelAttribute("comment") Comment comment, BoardVO boardVO,
 	    BindingResult bindingResult, ModelMap model, @RequestParam HashMap<String, String> map, HttpServletRequest request) throws Exception {
 		
 		LoginVO user = (LoginVO)EgovUserDetailsHelper.getAuthenticatedUser();
@@ -158,17 +174,30 @@ public class ArticleCommentController {
 		    return "forward:/cop/bbs/selectArticleDetail.do";
 		}
 	
-		if (isAuthenticated) {
-		    comment.setFrstRegisterId(user == null ? "" : EgovStringUtil.isNullToString(user.getUniqId()));
-		    comment.setWrterId(user == null ? "" : EgovStringUtil.isNullToString(user.getUniqId()));
-		    comment.setWrterNm(user == null ? "" : EgovStringUtil.isNullToString(user.getName()));
-		    
-		    
-		    ArticleCommentService.insertArticleComment(comment);
-		    
-		    commentVO.setCommentCn("");
-		    commentVO.setCommentNo("");
+		BoardMasterVO vo = new BoardMasterVO();
+
+		vo.setBbsId(boardVO.getBbsId());
+		vo.setUniqId((user == null || user.getUniqId() == null) ? "" : user.getUniqId());
+		BoardMasterVO master = BBSMasterService.selectBBSMasterInf(vo);
+
+		if(!master.getAuthComment().equals("GUE")) { // 접근 권한 비회원 이상(회원 & 관리자)			
+			if (!isAuthenticated ) { // 로그인이 안되어 있을 경우
+				return "forward:/uat/uia/LoginUsr.do";
+			}
+			if(!master.getAuthComment().equals(user.getUserSe()) && !user.getUserSe().equals("USR")) { 
+				return "egovframework/com/admin/cmm/error/accessDenied";
+			}
 		}
+		
+	    comment.setFrstRegisterId(user == null ? "" : EgovStringUtil.isNullToString(user.getUniqId()));
+	    comment.setWrterId(user == null ? commentVO.getWrterId() : EgovStringUtil.isNullToString(user.getUniqId()));
+	    comment.setWrterNm(user == null ? commentVO.getWrterNm() : EgovStringUtil.isNullToString(user.getName()));
+
+	    ArticleCommentService.insertArticleComment(comment);
+	    
+	    commentVO.setCommentCn("");
+	    commentVO.setCommentNo("");
+
 		
 		String chkBlog = map.get("blogAt");
 		
@@ -190,12 +219,34 @@ public class ArticleCommentController {
      * @throws Exception
      */
     @RequestMapping("/cop/cmt/deleteArticleComment.do")
-    public String deleteArticleComment(@ModelAttribute("searchVO") CommentVO commentVO, @ModelAttribute("comment") Comment comment, 
+    public String deleteArticleComment(@ModelAttribute("searchVO") CommentVO commentVO, @ModelAttribute("comment") Comment comment, BoardVO boardVO,
     		ModelMap model, @RequestParam HashMap<String, String> map, HttpServletRequest request) throws Exception {
 		@SuppressWarnings("unused")
 		LoginVO user = (LoginVO)EgovUserDetailsHelper.getAuthenticatedUser();
 		Boolean isAuthenticated = EgovUserDetailsHelper.isAuthenticated();
-	
+
+		BoardMasterVO vo = new BoardMasterVO();
+
+		vo.setBbsId(boardVO.getBbsId());
+		vo.setUniqId((user == null || user.getUniqId() == null) ? "" : user.getUniqId());
+		BoardMasterVO master = BBSMasterService.selectBBSMasterInf(vo);
+
+		if(!master.getAuthComment().equals("GUE")) { // 접근 권한 비회원 이상(회원 & 관리자)			
+			if (!isAuthenticated ) { // 로그인이 안되어 있을 경우
+				return "forward:/uat/uia/LoginUsr.do";
+			}
+			if(!master.getAuthComment().equals(user.getUserSe()) && !user.getUserSe().equals("USR")) { 
+				return "egovframework/com/admin/cmm/error/accessDenied";
+			}
+		}
+		
+		if(user==null) {
+			System.out.println("확인 : 여기까진옴");
+			String password = request.getParameter("pass");
+			System.out.println("확인 : " + password);
+			return "egovframework/com/admin/cmm/error/accessDenied";
+		}
+		
 		if (isAuthenticated) {
 		    ArticleCommentService.deleteArticleComment(commentVO);
 		}
@@ -210,6 +261,16 @@ public class ArticleCommentController {
 		}else{
 			return "redirect:"+request.getHeader("referer");
 		}
+    }
+    
+    /**
+     * [추가]-2021.06.01
+     * 비회원이 댓글을 삭제하기위한 페이지를 이동하기 위해
+     *
+     */
+    @RequestMapping("/cop/cmt/deleteArticleCommentPre.do")
+    public String deleteArticleCommentPre() {
+    	return "egovframework/com/web/board/passwordCheck";
     }
     
     
@@ -263,7 +324,7 @@ public class ArticleCommentController {
 		model.addAttribute("articleCommentVO", articleCommentVO);
 		
 		
-		return "egovframework/com/admin/cop/cmt/AdminArticleCommentList";
+		return "egovframework/com/admin/cop/cmt/ArticleCommentList";
     }
     
     
@@ -278,35 +339,46 @@ public class ArticleCommentController {
      * @throws Exception
      */
     @RequestMapping("/cop/cmt/updateArticleComment.do")
-    public String updateArticleComment(@ModelAttribute("searchVO") CommentVO commentVO, @ModelAttribute("comment") Comment comment, 
-	    BindingResult bindingResult, ModelMap model, HttpServletRequest request) throws Exception {
+    public String updateArticleComment(@ModelAttribute("searchVO") CommentVO commentVO, @ModelAttribute("comment") Comment comment,
+	    BindingResult bindingResult, ModelMap model, HttpServletRequest request, HttpServletResponse response) throws Exception {
     
 		LoginVO user = (LoginVO)EgovUserDetailsHelper.getAuthenticatedUser();
-		Boolean isAuthenticated = EgovUserDetailsHelper.isAuthenticated();
-	
+
 		beanValidator.validate(comment, bindingResult);
 		if (bindingResult.hasErrors()) {
 		    model.addAttribute("msg", "내용은 필수 입력 값입니다.");
 		    
 		    return "forward:/cop/bbs/selectArticleDetail.do";
 		}
-	
-		if (isAuthenticated) {
-		    comment.setLastUpdusrId(user == null ? "" : EgovStringUtil.isNullToString(user.getUniqId()));
-		    
-		    ArticleCommentService.updateArticleComment(comment);
-		    
-		    commentVO.setCommentCn("");
-		    commentVO.setCommentNo("");
+		
+		if(user==null) {	//로그인 안되어있을 때
+			String wrterId = comment.getWrterId();		// 이름
+			String wrterNm = comment.getWrterNm();		// 비밀번호
+			
+			CommentVO articleCommentVO = new CommentVO();
+			articleCommentVO = ArticleCommentService.selectArticleCommentDetail(commentVO);
+			
+			if(!articleCommentVO.getWrterId().equals(wrterId) || !articleCommentVO.getWrterNm().equals(wrterNm)) {
+				PrintWriter out = response.getWriter();
+				response.setContentType("text/html; charset=UTF-8");
+				out.print("<script lauguage='javascript'>");
+				out.print("alert('이름과 비밀번호가 일치하지 않습니다.');");
+				out.print("location.href='"+request.getHeader("referer")+"';");
+				out.print("</script>");
+				out.flush();
+				out.close();
+				
+				return "blank";
+			}
 		}
-		Enumeration<String> params = request.getParameterNames();
-		String strParam = "";
-		while(params.hasMoreElements()) {
-		    String name = (String)params.nextElement();
-		    String value = request.getParameter(name);
-		    strParam += name + "=" + value + "&";
-		}
-	
+		
+	    comment.setLastUpdusrId(user == null ? "" : EgovStringUtil.isNullToString(user.getUniqId()));
+	    
+	    ArticleCommentService.updateArticleComment(comment);
+	    
+	    commentVO.setCommentCn("");
+	    commentVO.setCommentNo("");
+
 		return "redirect:"+request.getHeader("referer");
     }	
 }
